hng13/backend/hng13-stage2/requirements.txt
fastapi
pydantic
uvicorn
SQLAlchemy
python-dotenv
requests
pillow
mysql-connector-python
mangum
